+++ packages/plugin-html/src/HtmlComponent.jsx
@@ -1,41 +1,84 @@
-import React from 'react';
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { mergeStyles, isValidUrl, normalizeUrl, validate } from 'wix-rich-content-common';
 
-import { SRC_TYPE_HTML, SRC_TYPE_URL, DEFAULT_COMPONENT_DATA } from './constants';
+import { SRC_TYPE_HTML, SRC_TYPE_URL, DEFAULT_COMPONENT_DATA, INIT_HEIGHT, INIT_WIDTH } from './constants';
 import schema from '../statics/data-schema.json';
 import IframeHtml from './IframeHtml';
 import IframeUrl from './IframeUrl';
 import htmlComponentStyles from '../statics/styles/HtmlComponent.scss';
 
-const HtmlComponent = props => {
-  const styles = mergeStyles({ styles: htmlComponentStyles, theme: props.theme });
-  validate(props.componentData, schema);
-  const { blockProps, componentData: { src, srcType, config: { width, height } = {} } } = props;
-  const style = { width, height };
-  const readOnly = blockProps ? blockProps.readOnly : true;
+class HtmlComponent extends Component {
+  constructor(props) {
+    super(props);
+    this.styles = mergeStyles({ styles: htmlComponentStyles, theme: props.theme });
+  }
 
-  return (
-    <div className={styles.htmlComponent} style={style} data-hook="HtmlComponent">
-      {srcType === SRC_TYPE_HTML && src && (
-        <IframeHtml
-          key={SRC_TYPE_HTML}
-          tabIndex={readOnly ? -1 : 0}
-          html={src}
-          src={props.settings.htmlIframeSrc}
-        />
-      )}
+  componentDidMount() {
+    if (!this.props.componentData.config.width) {
+      if (this.props.settings && this.props.settings.width) {
+        this.props.componentData.config.width = this.props.settings.width;
+      } else if (this.element) {
+        const { width } = this.element.getBoundingClientRect();
+        this.props.componentData.config.width = width;
+      } else {
+        this.props.componentData.config.width = INIT_WIDTH;
+      }
+    }
 
-      {srcType === SRC_TYPE_URL && isValidUrl(src) && (
-        <IframeUrl key={SRC_TYPE_URL} tabIndex={readOnly ? -1 : 0} src={normalizeUrl(src)} />
-      )}
+    if (!this.props.componentData.config.height) {
+      if (this.props.settings && this.props.settings.height) {
+        this.props.componentData.config.height = this.props.settings.height;
+      } else {
+        this.props.componentData.config.height = INIT_HEIGHT;
+      }
+    }
+  }
 
-      {!src && !isValidUrl(src) && (
-        <div className={styles.htmlComponent_placeholder} />
-      )}
-    </div>
-  );
-};
+  render() {
+    const { styles, props } = this;
+    validate(props.componentData, schema);
+    const {
+      blockProps,
+      componentData: {
+        src,
+        srcType,
+        config: {
+          width: currentWidth,
+          height: currentHeight
+        } = {}
+      },
+      settings: {
+        htmlIframeSrc,
+        width,
+        height
+      } = {} } = props;
+
+    const style = { width: currentWidth || width || INIT_WIDTH, height: currentHeight || height || INIT_HEIGHT };
+    const readOnly = blockProps ? blockProps.readOnly : true;
+
+    return (
+      <div className={styles.htmlComponent} ref={ref => this.element = ref} style={style} data-hook="HtmlComponent">
+        {srcType === SRC_TYPE_HTML && src && (
+          <IframeHtml
+            key={SRC_TYPE_HTML}
+            tabIndex={readOnly ? -1 : 0}
+            html={src}
+            src={htmlIframeSrc}
+          />
+        )}
+
+        {srcType === SRC_TYPE_URL && isValidUrl(src) && (
+          <IframeUrl key={SRC_TYPE_URL} tabIndex={readOnly ? -1 : 0} src={normalizeUrl(src)} />
+        )}
+
+        {!src && !isValidUrl(src) && (
+          <div className={styles.htmlComponent_placeholder} />
+        )}
+      </div>
+    );
+  }
+}
 
 HtmlComponent.propTypes = {
   componentData: PropTypes.object.isRequired,
@@ -44,6 +87,12 @@ HtmlComponent.propTypes = {
   theme: PropTypes.object,
   settings: PropTypes.shape({
     htmlIframeSrc: PropTypes.string.isRequired,
+    width: PropTypes.number,
+    minWidth: PropTypes.number,
+    maxWidth: PropTypes.number,
+    height: PropTypes.number,
+    minHeight: PropTypes.number,
+    maxHeight: PropTypes.number,
   }).isRequired,
 };