+++ packages/plugin-html/src/constants.js
@@ -4,9 +4,11 @@ export const SRC_TYPE_URL = 'url';
 export const MAX_ALIGNMENT_WIDTH = 739;
 
 export const MIN_WIDTH = 35;
+export const INIT_WIDTH = 740;
 export const MAX_WIDTH = 940;
 
 export const MIN_HEIGHT = 35;
+export const INIT_HEIGHT = 242;
 export const MAX_HEIGHT = 1200;
 export const MAX_HEIGHT_INPUT = 9999;
 
@@ -15,7 +17,5 @@ export const DEFAULT_COMPONENT_DATA = {
   src: '',
   config: {
     alignment: 'center',
-    width: 740,
-    height: 242,
   },
 };